+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Eng #103: Replace `RuntimeException` to `LogicException` when file name callback of `FileMailer` returns non-string
   value and improve exception message (@vjik)
 - New #104: Add `FileMailer` configuration (@vjik)
+- Chg #105: Change result type of `MessageInterface::with*` methods from `self` to `static` (@vjik)
 
 ## 5.1.0 July 02, 2024